# pkg/handlers/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "sigma",
        "contact": {
            "name": "sigma",
            "url": "https://github.com/go-sigma/sigma"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "1.3.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/namespaces/{namespace_id}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Namespace"
                ],
                "summary": "Get namespace",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Namespace id",
                        "name": "namespace_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.NamespaceItem"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Namespace"
                ],
                "summary": "Update namespace",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Namespace id",
                        "name": "namespace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Namespace object",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateNamespaceRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Namespace"
                ],
                "summary": "Delete namespace",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Namespace id",
                        "name": "namespace_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    }
                }
            }
        },
...
        "/{provider}/repos/coderepos/{id}/branches/{name}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CodeRepository"
                ],
                "summary": "Get specific name code repository branch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "code repository provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Code repository id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Branch name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.CodeRepositoryBranchItem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "enums.Visibility": {
            "type": "string",
            "enum": [
                "private",
                "public"
            ],
            "x-enum-varnames": [
                "VisibilityPrivate",
                "VisibilityPublic"
            ]
        },
...
        "errcode.ErrCode": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code provides a unique, string key, often capitalized with\nunderscores, to identify the error code. This value is used as the\nkeyed value when serializing api errors.",
                    "type": "string",
                    "example": "UNAUTHORIZED"
                },
                "description": {
                    "description": "Description provides a complete account of the errors purpose, suitable\nfor use in documentation.",
                    "type": "string",
                    "example": "The access controller was unable to authenticate the client. Often this will be accompanied by a Www-Authenticate HTTP response header indicating how to authenticate."
                },
                "http_status_code": {
                    "description": "HTTPStatusCode provides the http status code that is associated with\nthis error condition.",
                    "type": "integer",
                    "example": 401
                },
                "title": {
                    "description": "Title is a short, human readable description of the error condition\nincluded in API responses.",
                    "type": "string",
                    "example": "Authentication required"
                }
            }
        },
...
        "types.CodeRepositoryBranchItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "main"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                }
            }
        },
...
        "types.NamespaceItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                },
                "description": {
                    "type": "string",
                    "example": "i am just description"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "test"
                },
                "overview": {
                    "type": "string",
                    "example": "i am just overview"
                },
                "repository_count": {
                    "type": "integer",
                    "example": 10
                },
                "repository_limit": {
                    "type": "integer",
                    "example": 10
                },
                "size": {
                    "type": "integer",
                    "example": 10000
                },
                "size_limit": {
                    "type": "integer",
                    "example": 10000
                },
                "tag_count": {
                    "type": "integer",
                    "example": 10
                },
                "tag_limit": {
                    "type": "integer",
                    "example": 10
                },
                "updated_at": {
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                },
                "visibility": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.Visibility"
                        }
                    ],
                    "example": "private"
                }
            }
        },
...
        "types.UpdateNamespaceRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 30,
                    "example": "i am just description"
                },
                "overview": {
                    "type": "string",
                    "maxLength": 100000,
                    "example": "i am just overview"
                },
                "repository_limit": {
                    "type": "integer",
                    "example": 10000
                },
                "size_limit": {
                    "type": "integer",
                    "example": 10000
                },
                "tag_limit": {
                    "type": "integer",
                    "example": 10000
                },
                "visibility": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.Visibility"
                        }
                    ],
                    "example": "public"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}